Load TAG_Lists with TAG_End for contents.
But only if they're empty. The returned object looks like an empty list,
but it's read-only because we don't know the tag-type of the contents.
(Well, we do, it's TAG_End, but we assume a list of TAG_End is empty
because that's the only way this even starts to make sense.) We don't
want to make any assumptions, so you might need to check for these and
just replace them entirely with a TAG_List that makes sense.

Truthfully, I think TAG_Lists of TAG_Ends are a mistake and a bug, and
it pains me that Minecraft's NBT code is even capable of writing such a
thing.